Split game/mod.rs into actions, economy, and turn modules
Pure moves — no behavior changes. mod.rs (2,314 lines) kept growing as
every track landed in it; it now retains the Game struct, constants,
RNG-stream derivation, initialize_player, save/load, and the test suite
(1,027 lines). The moved code:

- game/actions.rs: the action_* handlers, their quality/song-selection
  helpers, get_sorted_regions, live_fame_cap, and execute_action.
- game/economy.rs: recording/pressing costs, sales scoring,
  distribution, release outcome, and the weekly release/catalog payout.
- game/turn.rs: process_turn, weekly events and news, offer aging,
  visibility decay, random/historical event application, and the
  game-over checks.

Mechanical adjustments only, called out for review:
- Methods called across module boundaries (or from the tests, which sit
  in game::tests) went fn -> pub(super); nothing gained full pub.
- Four stale '// --- ... (Step N) ---' section markers and the
  'calculate_royalties is removed' note were dropped, replaced by the
  module doc headers.
- rustfmt re-wrapped two signatures that pub(super) pushed past the
  line limit.

Verified: the test list is byte-identical before and after (43 tests);
cargo test, clippy --all-targets, and fmt --check are all clean.
